requirements.txt
mss==9.0.1
pillow==10.2.0
requests==2.31.0
pyyaml==6.0.1
fastapi==0.109.0
uvicorn==0.27.0
ollama==0.1.6